+++ arch/loongarch/kernel/module.c
@@ -18,7 +18,6 @@
 #include <linux/ftrace.h>
 #include <linux/string.h>
 #include <linux/kernel.h>
-#include <linux/execmem.h>
 #include <asm/alternative.h>
 #include <asm/inst.h>
 #include <asm/unwind.h>
@@ -489,23 +488,6 @@ int apply_relocate_add(Elf_Shdr *sechdrs, const char *strtab,
 	}
 
 	return 0;
-}
-
-static struct execmem_info execmem_info __ro_after_init = {
-	.ranges = {
-		[EXECMEM_DEFAULT] = {
-			.pgprot = PAGE_KERNEL,
-			.alignment = 1,
-		},
-	},
-};
-
-struct execmem_info __init *execmem_arch_setup(void)
-{
-	execmem_info.ranges[EXECMEM_DEFAULT].start = MODULES_VADDR;
-	execmem_info.ranges[EXECMEM_DEFAULT].end = MODULES_END;
-
-	return &execmem_info;
 }
 
 static void module_init_ftrace_plt(const Elf_Ehdr *hdr,
